requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.1
dnspython==2.2.1
frozenlist==1.3.1
humanize==4.4.0
idna==3.4
motor==3.0.0
multidict==6.0.2
py-cord==2.2.2
pymongo==4.3.2
python-dotenv==0.21.0
yarl==1.8.1
